\name{p3state.msm-package}
\alias{p3state.msm-package}
\alias{p3state.msm}
\docType{package}
\title{Inference in an Illness-Death Model}
\description{
p3state.msm  provides functions for estimating semi-parametric regression models but also to implement nonparametric estimators for the transition probabilities. The methods can also be used in progressive three-state models.

In progressive three-state models, estimators for other quantities such as the bivariate distribution function (for the sequentially ordered events) are also given
}
%\details{
%\tabular{ll}{
%Package: \tab p3state.msm\cr
%Type: \tab Package\cr
%Version: \tab 1.0\cr
%Date: \tab 2009-06-07\cr
%License: \tab 2.8.1\cr
%LazyLoad: \tab yes\cr
%}
%~~ An overview of how to use the package, including the most important ~~
%~~ functions ~~
%}
\author{
Luis Meira-Machado and Javier Roca Pardinas \email{roca@uvigo.es}

Maintainer: Luis Meira-Machado \email{lmachado@mct.uminho.pt}
}

%\references{
%~~ Literature or other references for background information ~~
%}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
%~~ the R documentation directory ~~
%\keyword{ package }
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%\examples{
%~~ simple examples of the most important functions ~~
%}
